Finish torus, refactor p3 to float

+++ shapes/cone.h
@@ -6,7 +6,7 @@
 class Cone : public OpenGLShape
 {
 public:
-    Cone(int p1, int p2, int p3);
+    Cone(int p1, int p2, float p3);
 
     virtual void reCalculateVertices();
 };
