+++ shapes/cube.h
@@ -6,7 +6,7 @@
 class Cube : public OpenGLShape
 {
 public:
-    Cube(int p1, int p2, int p3);
+    Cube(int p1, int p2, float p3);
 
     virtual void reCalculateVertices();
 };
